fix: OpenApiService 리팩토링
